requirements.txt
backports.shutil-get-terminal-size==1.0.0
decorator==4.2.1
Django==1.11.24
django-scss==0.6.1
django-static-precompiler==1.8.2
djangorestframework==3.8.2
enum34==1.1.6
ipython==5.6.0
ipython-genutils==0.2.0
pathlib2==2.3.0
pexpect==4.4.0
pickleshare==0.7.4
prompt-toolkit==1.0.15
ptyprocess==0.5.2
Pygments==2.2.0
pytz==2018.4
scandir==1.7
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
typing==3.6.4
wcwidth==0.1.7
